core: fix 32-bit TLV decoding and extend GTPv1 fuzzing
Pass short GTPv1 inputs to ogs_gtp1_parse_msg() so the fuzzer also
exercises header-length validation. Rename the GTPv2 message variable
to distinguish it from the GTPv1 counterpart.

The fuzzer exposed undefined behavior in 32-bit TLV decoding when the
first octet has its high bit set. It was promoted to int before being
shifted by 24, which can overflow a signed int.

Cast the first octet to uint32_t before shifting in both decoding paths
and add regression coverage for high-bit values.

Related PR: #4751

+++ lib/core/ogs-tlv-msg.c
@@ -547,7 +547,7 @@ static int tlv_parse_leaf(void *msg, ogs_tlv_desc_t *desc, ogs_tlv_t *tlv)
             ogs_error("Invalid TLV length %d. It should be 4", tlv->length);
             return OGS_ERROR;
         }
-        v->u32 = ((((uint8_t*)tlv->value)[0]<<24)&0xff000000) |
+        v->u32 = (((uint32_t)((uint8_t*)tlv->value)[0]<<24)&0xff000000) |
                ((((uint8_t*)tlv->value)[1]<<16)&0x00ff0000) |
                ((((uint8_t*)tlv->value)[2]<< 8)&0x0000ff00) |
                ((((uint8_t*)tlv->value)[3]    )&0x000000ff);
